pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "folioblog"
version = "1.0.0"
description = "A PortFolio/Blog website build with the CMS Wagtail"
authors = ["Yannick Chabbert <yannick.chabbert.it@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
wagtail = "^4.0.4"
django-social-share = "^2.3.0"
django-compressor = "^4.1"
redis = "^4.3.4"
hiredis = "^2.0.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
tblib = "^1.7.0"
factory-boy = "^3.2.1"
wagtail-factories = "^3.1.0"
requests-mock = "^1.10.0"
selenium = "^4.5.0"
django-extensions = "^3.2.1"
django-debug-toolbar = "^3.7.0"
pre-commit = "^2.20.0"
tox = "^3.26.0"
coverage = {extras = ["toml"], version = "^6.5.0"}
pylama = {extras = ["toml"], version = "^8.4.1"}

[tool.pylama]
skip = "*/migrations/*,*/settings/*,*/static/*"
linters = "pycodestyle,pyflakes,mccabe"

[tool.pylama.linter.pycodestyle]
max_line_length = 119

[tool.pylama.linter.mccabe]
complexity = 15

[tool.isort]
combine_as_imports = true
multi_line_output = 5
include_trailing_comma = true
skip = "migrations,docs,media,static"
sections = "FUTURE,STDLIB,DJANGO,WAGTAIL,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_django = "django"
known_wagtail = "wagtail"
known_first_party = "folioblog"

[tool.coverage.run]
source = [ "folioblog" ]
branch = true
omit = [
    "manage.py",
    "folioblog/wsgi.py",
    "folioblog/core/compressor.py",
    "folioblog/settings/dev.py",
    "folioblog/settings/local.py",
    "folioblog/settings/prod.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "def __str__",
    "def __repr__",
]
